Exams/Python OOP Retake Exam - 22 Aug 2020/project/rooms/room.py
class Room:
    _DAYS_IN_MONTH = 30

    def __init__(self, family_name: str, budget: float, members_count: int):
        self.family_name = family_name
        self.budget = budget
        self.members_count = members_count
        self.children = []
        self.expenses = 0

    @property
    def expenses(self):
        return self.__expenses

    @expenses.setter
    def expenses(self, value):
        if value < 0:
            raise ValueError("Expenses cannot be negative")
        self.__expenses = value

    def calculate_expenses(self, *args):
        result = 0
        for arg in args:
            result += sum([el.cost for el in arg])
        self.expenses = result * Room._DAYS_IN_MONTH
